Add build script
This build script allows capng to be linked statically
to libcap-ng and/or select the library's location.
To set the link type the build script will check the
env. variable LIBCAPNG_LINK_TYPE, if its value is
'dylib' or 'static' the provided link type is selected, in
other case the `dylib` is selected.

LIBCAPNG_LIB_PATH=<path> will add the <path> directory
to the compiler's library search path (i.e., -L <path>

Signed-off-by: German Maglione <gmaglione@redhat.com>